requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.12.2
Flask==2.2.5
fsspec==2023.1.0
huggingface-hub==0.16.4
idna==3.6
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.21.6
packaging==23.2
Pillow==9.5.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
tokenizers==0.13.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.2
transformers==4.30.2
typing_extensions==4.7.1
urllib3==1.26.7
Werkzeug==2.2.3
zipp==3.15.0
gunicorn
